+++ project/LinkerdBuild.scala
@@ -91,7 +91,6 @@ object LinkerdBuild extends Base {
 
     val marathon = projectDir("namer/marathon")
       .dependsOn(LinkerdBuild.marathon, core)
-      .dependsOn(core)
       .withTests()
 
     val serversets = projectDir("namer/serversets")
@@ -99,14 +98,19 @@ object LinkerdBuild extends Base {
       .withTests()
       .dependsOn(core % "compile->compile;test->test")
 
+    val zkLeader = projectDir("namer/zk-leader")
+      .dependsOn(core)
+      .withTwitterLib(Deps.zkCandidate)
+
     val all = projectDir("namer")
-      .aggregate(core, consul, fs, k8s, marathon, serversets)
+      .aggregate(core, consul, fs, k8s, marathon, serversets, zkLeader)
   }
 
   val admin = projectDir("admin")
     .dependsOn(configCore, Namer.core)
     .withTwitterLib(Deps.twitterServer)
     .withTwitterLib(Deps.finagle("stats"))
+    .withTests()
 
   val ConfigFileRE = """^(.*)\.yaml$""".r
 
@@ -179,6 +183,7 @@ object LinkerdBuild extends Base {
       val interpreterThrift = projectDir("namerd/iface/interpreter-thrift")
         .dependsOn(core)
         .dependsOn(interpreterThriftIdl)
+        .withLib(Deps.guava)
         .withTwitterLibs(Deps.finagle("thrift"), Deps.finagle("thriftmux"))
         .withTests()
 
@@ -427,7 +432,7 @@ object LinkerdBuild extends Base {
       .withTwitterLib(Deps.finagle("stats") % Minimal)
       // Bundle is includes all of the supported features:
       .configDependsOn(Bundle)(
-        Namer.consul, Namer.k8s, Namer.marathon, Namer.serversets,
+        Namer.consul, Namer.k8s, Namer.marathon, Namer.serversets, Namer.zkLeader,
         Interpreter.namerd,
         Protocol.mux, Protocol.thrift,
         Tracer.zipkin,
@@ -485,6 +490,7 @@ object LinkerdBuild extends Base {
   val namerK8s = Namer.k8s
   val namerMarathon = Namer.marathon
   val namerServersets = Namer.serversets
+  val namerZkLeader = Namer.zkLeader
 
   val namerd = Namerd.all
   val namerdExamples = Namerd.examples