+++ CHANGES.md
@@ -1,5 +1,9 @@
 ## x.x.x
 
+* Add zkLeader namer to allow discovery of services through ZooKeeper leader
+  election.
+* Add HTTP path request identifier, which builds destinations from a
+  configurable number of HTTP URI path segments.
 * The path prefix `/#` now indicates that the path should be processed by a
   namer.  A namer matches a path starting with `/#/<prefix>`.